+++ src/UIScrollBar.cpp
@@ -481,7 +481,8 @@ auto WINAPI LongUI::UIScrollBarA::CreateControl(CreateEventType type, pugi::xml_
             point_list[1] = { BASIC_SIZE_NEAR , BASIC_SIZE_MID };
             point_list[2] = { BASIC_SIZE_MID , BASIC_SIZE_FAR };
             assert(!s_apArrowPathGeometry[UIScrollBarA::Arrow_Left]);
-            s_apArrowPathGeometry[UIScrollBarA::Arrow_Left] = create_geo(point_list, lengthof(point_list));
+            s_apArrowPathGeometry[UIScrollBarA::Arrow_Left] = 
+                create_geo(point_list, static_cast<uint32_t>(lengthof(point_list)));
         }
         // TOP 上箭头
         {
@@ -489,7 +490,8 @@ auto WINAPI LongUI::UIScrollBarA::CreateControl(CreateEventType type, pugi::xml_
             point_list[1] = { BASIC_SIZE_MID, BASIC_SIZE_NEAR };
             point_list[2] = { BASIC_SIZE_FAR, BASIC_SIZE_MID };
             assert(!s_apArrowPathGeometry[UIScrollBarA::Arrow_Top]);
-            s_apArrowPathGeometry[UIScrollBarA::Arrow_Top] = create_geo(point_list, lengthof(point_list));
+            s_apArrowPathGeometry[UIScrollBarA::Arrow_Top] = 
+                create_geo(point_list, static_cast<uint32_t>(lengthof(point_list)));
         }
         // RIGHT 右箭头
         {
@@ -498,7 +500,8 @@ auto WINAPI LongUI::UIScrollBarA::CreateControl(CreateEventType type, pugi::xml_
             point_list[1] = { BASIC_SIZE_FAR , BASIC_SIZE_MID };
             point_list[2] = { BASIC_SIZE_MID , BASIC_SIZE_FAR };
             assert(!s_apArrowPathGeometry[UIScrollBarA::Arrow_Right]);
-            s_apArrowPathGeometry[UIScrollBarA::Arrow_Right] = create_geo(point_list, lengthof(point_list));
+            s_apArrowPathGeometry[UIScrollBarA::Arrow_Right] = 
+                create_geo(point_list, static_cast<uint32_t>(lengthof(point_list)));
         }
         // BOTTOM 下箭头
         {
@@ -506,7 +509,8 @@ auto WINAPI LongUI::UIScrollBarA::CreateControl(CreateEventType type, pugi::xml_
             point_list[1] = { BASIC_SIZE_MID, BASIC_SIZE_FAR };
             point_list[2] = { BASIC_SIZE_FAR, BASIC_SIZE_MID };
             assert(!s_apArrowPathGeometry[UIScrollBarA::Arrow_Bottom]);
-            s_apArrowPathGeometry[UIScrollBarA::Arrow_Bottom] = create_geo(point_list, lengthof(point_list));
+            s_apArrowPathGeometry[UIScrollBarA::Arrow_Bottom] = 
+                create_geo(point_list, static_cast<uint32_t>(lengthof(point_list)));
         }
     }
         break;
